Almost finalized product

Approved-by: William Summy

+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/UserDao.java
@@ -10,7 +10,7 @@ public interface UserDao {
 
     User findByUsername(String username);
 
-    int findIdByUsername(String username);
+    Long findIdByUsername(String username);
 
     boolean create(String username, String password);
 }
